+++ src/main/java/subway/line/LineDao.java
@@ -6,20 +6,16 @@ import org.springframework.jdbc.core.RowMapper;
 import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.stereotype.Repository;
-import org.springframework.util.ReflectionUtils;
-import subway.DuplicateException;
-import subway.section.Section;
-import subway.station.Station;
-import subway.station.StationDao;
+import subway.exception.DuplicateException;
 
-import java.lang.reflect.Field;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.sql.Statement;
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
+
+import static subway.line.LineQuery.*;
 
 @Repository
 public class LineDao {
@@ -33,17 +29,14 @@ public class LineDao {
         if (hasDuplicateName(line.getName())) {
             throw new DuplicateException();
         }
-
         return insertAtDB(line);
     }
 
     public void update(Line line) {
-        String updateQuery = "update line set name = ?, color = ? where id = ?";
         jdbcTemplate.update(updateQuery, line.getName(), line.getColor(), line.getId());
     }
 
     public Optional<Line> findById(Long id) {
-        String selectByIdQuery = "select * from line where id = ?";
         try {
             return Optional.ofNullable(jdbcTemplate.queryForObject(selectByIdQuery, new LineMapper(), id));
         } catch (EmptyResultDataAccessException e) {
@@ -52,18 +45,15 @@ public class LineDao {
     }
 
     public List<Line> findAll() {
-        String selectAllQuery = "select * from line";
         List<Line> lines = jdbcTemplate.query(selectAllQuery, new LineMapper());
         return lines;
     }
 
     public void deleteById(Long id) {
-        String deleteByIdQuery = "delete from line where id = ?";
         jdbcTemplate.update(deleteByIdQuery, id);
     }
 
     public boolean hasDuplicateName(String name) {
-        String countByNameQuery = "select count(*) from line where name = ?";
         return jdbcTemplate.queryForObject(countByNameQuery, int.class, name) != 0;
     }
 
@@ -72,7 +62,7 @@ public class LineDao {
 
         jdbcTemplate.update(con -> {
             PreparedStatement psmt = con.prepareStatement(
-                    "insert into line (name, color) values(?, ?)",
+                    insertQuery,
                     Statement.RETURN_GENERATED_KEYS);
             psmt.setString(1, line.getName());
             psmt.setString(2, line.getColor());
